Refactor to use the ScalaJS `crossBuild` plugin
