+++ index.html
@@ -12,9 +12,8 @@
     <canvas style="display: block; margin-left: auto; margin-right: auto;" id="canvas" width="255" height="255"/>
 </div>
 
-<script type="text/javascript" src="./core/target/scala-2.11/doodle-core-fastopt.js"></script>
-<script type="text/javascript" src="./canvas/target/scala-2.11/doodle-canvas-fastopt.js"></script>
-<script type="text/javascript" src="./canvas/target/scala-2.11/doodle-canvas-launcher.js"></script>
+<script type="text/javascript" src="./js/target/scala-2.11/doodle-fastopt.js"></script>
+<script type="text/javascript" src="./js/target/scala-2.11/doodle-launcher.js"></script>
 <script type="text/javascript" src="/workbench.js"></script>
 </body>
 </html>
