--- java2d/src/main/scala/Main.scala
@@ -1,3 +1,0 @@
-object Main extends App {
-  doodle.java2d.Draw(doodle.Example.picture)
-}
