+++ js/src/main/scala/doodle/js/Draw.scala
@@ -1,4 +1,4 @@
-package doodle.canvas
+package doodle.js
 
 import org.scalajs.dom
