+++ jvm/src/main/scala/Main.scala
@@ -1,0 +1,3 @@
+object Main extends App {
+  doodle.jvm.Draw(doodle.Example.picture)
+}
