+++ shared/src/main/scala/doodle/Example.scala
