+++ shared/src/main/scala/doodle/core/Angle.scala
