+++ shared/src/main/scala/doodle/core/BoundingBox.scala
