+++ shared/src/main/scala/doodle/core/Color.scala
