+++ shared/src/main/scala/doodle/core/DrawingContext.scala
