+++ shared/src/main/scala/doodle/core/Image.scala
