+++ shared/src/main/scala/doodle/core/UnsignedByte.scala
