+++ shared/src/test/scala/doodle/core/ColourSpec.scala
